+++ app/__init__.py
@@ -6,7 +6,7 @@
 # CreateTime : 2017-03-27 15:43
 # ===================================
 
-from flask import Flask, render_template
+from flask import Flask
 from flask_mail import Mail
 from flask_moment import Moment
 from flask_pagedown import PageDown
@@ -34,23 +34,18 @@ def create_app(config_name):
     login_manager.init_app(app)
     pagedown.init_app(app)
 
-    from .core.resources import main as main_blueprint
-    app.register_blueprint(main_blueprint)
+    from .resources import resource_blueprint
+    # app.register_blueprint(resource_blueprint, static_folder='static', template_folder='templates')
+    app.register_blueprint(resource_blueprint)
 
     from .core.auth import auth as auth_blueprint
     app.register_blueprint(auth_blueprint, url_prefix = '/auth')
 
-    from .core.resources import manage as manage_blueprint
+    from .core.manage import manage_blueprint
     app.register_blueprint(manage_blueprint, url_prefix = '/manage')
-    #
-    # from .api_1_0 import api as api_1_0_blueprint
-    # app.register_blueprint(api_1_0_blueprint, url_prefix='/api/v1.0')
+
+    from .api import api_v1_blueprint
+    app.register_blueprint(api_v1_blueprint, url_prefix='/api/v1.0')
 
 
     return app
-
-class MySQLAlchemy(SQLAlchemy):
-    def init_database(self, **kwargs):
-        database = kwargs.pop('database', None)
-
-        self.session.add()
