+++ app/core/auth/__init__.py
@@ -3,12 +3,12 @@
 # ScriptName : __init__.py.py
 # Author     : WFJ
 # Email      : wfj_sc@163.com
-# CreateTime : 2017-03-28 9:30
+# CreateTime : 2017-04-05 9:59
 # ===================================
+
 
 from flask import Blueprint
 
 auth = Blueprint('auth', __name__)
 
 from . import views
-
