+++ app/core/common/user_mixin.py
@@ -19,7 +19,6 @@ class UserMixin(object):
 
     @property
     def is_authenticated(self):
-        print "haha ------------"
         return True
 
     @property
