+++ app/core/manage/resource_views.py
