--- app/core/resources/manage/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-Author:         WFJ
-Version:        0.1.0
-FileName:       __init__.py.py
-CreateTime:     2017-04-04 19:46
-"""
-
