--- app/core/resources/manage/views.py
@@ -1,14 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-Author:         WFJ
-Version:        0.1.0
-FileName:       views.py
-CreateTime:     2017-04-04 19:47
-"""
-
-from app.utils.email import send_email
-from app.core.db.sql.models import User
-from flask import render_template, redirect, request, url_for, flash
-from flask_login import login_user, login_required, logout_user, current_user
