--- app/core/resources/posts/__init__.py
@@ -1,14 +1,0 @@
-# -*- coding: utf-8 -*-
-# ===================================
-# ScriptName : __init__.py.py
-# Author     : WFJ
-# Email      : wfj_sc@163.com
-# CreateTime : 2017-03-30 9:15
-# ===================================
-
-
-
-
-if __name__ == '__main__':
-    pass
-
