+++ app/resources/__init__.py
@@ -6,7 +6,10 @@
 # CreateTime : 2017-03-27 15:45
 # ===================================
 
-
 from flask import Blueprint
 
-api_v1_blueprint = Blueprint('v1', __name__)
+resource_blueprint = Blueprint('resource', __name__)
+# 这个地方很奇怪，为什么必须导出来？，还有模板和静态文件的路径为什么是app目录下
+from . import views
+from .logs import views
+from .persons import views
