+++ app/templates/base.html
@@ -82,7 +82,7 @@
                 <span class="icon-bar"></span>
                 <span class="icon-bar"></span>
             </button>
-            <a class="navbar-brand" href="{{ url_for('main.index') }}">index</a>
+            <a class="navbar-brand" href="{{ url_for('resource.index') }}">index</a>
         </div>
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav">
@@ -101,13 +101,13 @@
                             <img src="{{ current_user.gravatar(size=18) }}"> {{ current_user.name }} <b class="caret"></b>
                         </a>
                         <ul class="dropdown-menu">
-                            <li><a href="{{ url_for('main.user', name=current_user.name) }}">Personal Homepage</a></li>
+                            <li><a href="{{ url_for('resource.user', name=current_user.name) }}">Personal Homepage</a></li>
                             <li><a href="{{ url_for('auth.change_password')}}">Change Password</a></li>
                             <li><a href="{{ url_for('auth.change_email_request')}}">Change Email</a></li>
                             <li><a href="{{ url_for('auth.logout') }}">Logout</a></li>
                         </ul>
                     </li>
-                    <li><a href="{{ url_for('main.logs', name=current_user.name) }}">Logs</a></li>
+                    <li><a href="{{ url_for('resource.logs', name=current_user.name) }}">Logs</a></li>
                     {% if current_user.has_right('manage', 'show') %}
                         <li class="dropdown">
                             <a href='#' class="dropdown-toggle" data-toggle="dropdown">
