+++ app/templates/resources/log/log.html
